refactor : Seat와 중복되는 Ui 용 좌표 Model 제거
