+++ app/src/main/java/woowacourse/movie/domain/model/seat/Column.kt
@@ -1,0 +1,4 @@
+package woowacourse.movie.domain.model.seat
+
+@JvmInline
+value class Column(val value: Int)
