+++ app/src/main/java/woowacourse/movie/domain/model/seat/Row.kt
@@ -1,0 +1,4 @@
+package woowacourse.movie.domain.model.seat
+
+@JvmInline
+value class Row(val value: Int)
