+++ app/src/main/java/woowacourse/movie/view/complete/BookingCompleteActivity.kt
@@ -38,7 +38,6 @@ class BookingCompleteActivity : AppCompatActivity(), BookingCompleteContract.Vie
             insets
         }
         supportActionBar?.setDisplayHomeAsUpEnabled(true)
-        presenter.loadTicket()
     }
 
     override fun showTicket(ticket: Ticket) {
@@ -84,8 +83,8 @@ class BookingCompleteActivity : AppCompatActivity(), BookingCompleteContract.Vie
 
     private fun seatToLabel(seats: Set<Seat>): String {
         return seats.joinToString {
-            val rowLetter = ('A' + it.x - 1)
-            val columnNumber = it.y
+            val rowLetter = ('A' + it.x.value - 1)
+            val columnNumber = it.y.value
             "$rowLetter$columnNumber"
         }
     }
