+++ app/src/main/java/woowacourse/movie/view/seat/SeatView.kt
@@ -5,13 +5,10 @@ import android.widget.TableRow
 import androidx.core.view.children
 import woowacourse.movie.domain.model.seat.Seat
 import woowacourse.movie.view.seat.SeatRow
-import woowacourse.movie.view.seat.model.coord.Column
-import woowacourse.movie.view.seat.model.coord.Coordination
-import woowacourse.movie.view.seat.model.coord.Row
 
 class SeatView(
     private val seatTable: TableLayout,
-    private val onSeatClick: (Coordination) -> Unit,
+    private val onSeatClick: (Seat) -> Unit,
 ) {
     private val rows = mutableListOf<SeatRow>()
 
@@ -26,7 +23,6 @@ class SeatView(
     }
 
     fun updateSeats(selectedSeats: Set<Seat>) {
-        val selectedCoords = selectedSeats.map { Coordination(Column(it.x), Row(it.y)) }.toSet()
-        rows.forEach { it.updateSeats(selectedCoords) }
+        rows.forEach { it.updateSeats(selectedSeats) }
     }
 }
