--- app/src/main/java/woowacourse/movie/view/seat/model/coord/Column.kt
@@ -1,4 +1,0 @@
-package woowacourse.movie.view.seat.model.coord
-
-@JvmInline
-value class Column(val value: Int)
