--- app/src/main/java/woowacourse/movie/view/seat/model/coord/Coordination.kt
@@ -1,6 +1,0 @@
-package woowacourse.movie.view.seat.model.coord
-
-data class Coordination(
-    val x: Column,
-    val y: Row,
-)
