--- app/src/main/java/woowacourse/movie/view/seat/model/coord/Row.kt
@@ -1,4 +1,0 @@
-package woowacourse.movie.view.seat.model.coord
-
-@JvmInline
-value class Row(val value: Int)
