--- app/src/test/java/woowacourse/movie/domain/fixture/CoordinationFixture.kt
@@ -1,8 +1,0 @@
-package woowacourse.movie.domain.fixture
-
-import woowacourse.movie.view.seat.model.coord.Column
-import woowacourse.movie.view.seat.model.coord.Coordination
-import woowacourse.movie.view.seat.model.coord.Row
-
-val oneByOneCoord = Coordination(Column(1), Row(1))
-val twoByThreeCoord = Coordination(Column(2), Row(3))
