+++ app/src/test/java/woowacourse/movie/domain/fixture/SeatFixture.kt
@@ -1,8 +1,10 @@
 package woowacourse.movie.domain.fixture
 
+import woowacourse.movie.domain.model.seat.Column
+import woowacourse.movie.domain.model.seat.Row
 import woowacourse.movie.domain.model.seat.Seat
 
-val oneByOneSeat = Seat(1, 1)
+val oneByOneSeat = Seat(Column(1), Row(1))
 
-val oneByTowSeat = Seat(1, 12)
-val twoByThreeSeat = Seat(2, 3)
+val oneByTowSeat = Seat(Column(1), Row(12))
+val twoByThreeSeat = Seat(Column(2), Row(3))
